initial commit of pre-commit configuration
To use the pre-commit hooks install `pre-commit` with the following
instructions:

  https://pre-commit.com/#install

then run ONCE in your CP2K clone:

    pre-commit install

After that - before every commit - all the scripts configured in
`.pre-commit-config.yaml` will run on the files added in that
commit.

If any file changed (for example because they were modified by the
`prettify` script), the commit will be interrupted for you to
investigate and adding the prettyfied file to the commit if
necessary.