m_cpuid_static (C impl.): there seems to be no way to determine target flags in Fortran code (#331)
* Similar to other Fortran compilers, `gfortran -E -dM -mavx - < /dev/null | grep AVX` defines a variety of predefined macros (also similar to C).
* However, with a Fortran translation unit only a subset of these definitions disappears (`gfortran -E -dM -mavx my.F | grep AVX`).

To demonstrate and use `m_cpuid_static()` and `m_cpuid()`, a new (useful) CPHINT is introduced: "_the compiler target flags used to build this binary are insufficiently exploiting the extensions which are available for this CPU model_". The hint is shown after presenting the "CPU model name" (global environment output during CP2K's startup).

CPHINT is similar to CPWARN but does not increment the warning counter.