Cubic scaling RPA/GW/SOS-MP2: remove legacy code
This removes all code that has been replaced with a new implementation
based on DBCSR tensors. The following features have not been adapted to
the new implementation and are removed:
- Modified Atomic Orbitals (MAOs)
- full k-points for GW

The input section IM_TIME has been simplified and the number of
keywords/parameters has been reduced. Input for the RI metric has
been moved to a section RI_METRIC.