LS: Add direct solver for sign method on submatrices
Since the submatrices are dense and we require the entire spectrum to
calculate the sign function, calculating the sign function using a
direct eigendecomposition is competitive to iterative methods.

This method can be used by setting SUBMATRIX_SIGN_METHOD to DIRECT. The
setting defaults to NEWTONSCHULZ. To perform the eigendecomposition, we
need a symmetric matrix. Therefore this method can only be used when
SIGN_SYMMETRIC is set as well.

Co-authored-by: Robert Schade <robert.schade@uni-paderborn.de>