Bump LibXC to version 5.2.0 (#1899)
* Bump LibXC to version 8.2.0

* Fix redefined i_ref

The behavior of the underlying routine changed such that xc_f03_func_info_get_references does not increase the counter if the next reference is not found. Before, this routine returned -1 instead.

* Use LDA_X_ERF correctly

Before LibXC 5.2, there was a misleading description of the functional. Because these tests implement range-separated MP2, we should adjust the regtests according to the revised documentation of LDA_X_ERF.