Require arrays to be contiguous wherever possible
In case of non-blocking communication, we check contiguity with is_contiguous to prevent accidental copies from caller-site. If we were to call a non-blocking routine with a non-contiguous array and enforce contiguity in the interface, the compiler would make the array contiguous before calling the routine and copy it back afterwards such that MPI would not copy the result into the actual array and might even throw a segfault because the contiguous copy of the non-contiguous array might have already been cleaned up.