Fix detection of system Open MPI (#2578)
The commands mpirun and mpiexec employ different runtime environments. The search string to determine the MPI flavor is based on the this very name in the case of OpenMPI. When switching to mpiexec, it was forgotten to adjust it accordingly. This issue is absent in case of IntelMPI because mpirun is a apparently a mere alias of mpiexec. I do not know about the respective differences in the case of MPICH which is not available on our cluster but mistakes here are not important anyways because the toolchain defaults to MPICH if the system MPI is neither IntelMPI nor Open MPI.