Ignore SCF convergence failure for H2O benchmark
The old H2O benchmark files do not use an OUTER_SCF with OT which causes
that the first SCF iteration, e.g. already for H2O-32.inp, does not converge.
This convergence failure, however, causes now an program abort. The keyword
IGNORE_CONVERGENCE_FAILURE is added in order to recover the old behavior
and to keep the obtained timings comparable.