Sort items in code generated using fypp
When doing package builds, the result would be not reproducible, with
differences in a few files following the same pattern:

   MODULE list_callstackentry
  -   USE timings_base_type, ONLY: callstack_entry_type, routine_stat_type, routine_report_type
  +   USE timings_base_type, ONLY: routine_stat_type, callstack_entry_type, routine_report_type

It seems that this comes from the use of a Python set to deduplicate items. The
order of items in a set is not stable, it is based on the hashes, which are
seeded randomly for each invocation of Python.

While at it, do not create an unnecessary list and do not use '_' as a variable
name. ('_' is used when the value is supposed to be ignored.)